moved upsampling code in separate folder, so that it can be pip installed
