+++ upsampling/setup.py
@@ -1,0 +1,7 @@
+from setuptools import setup
+
+setup(
+    name='upsampler',
+    package_dir={'':'src'},
+    packages=['upsampler']
+    )
