+++ upsampling/src/upsampler/__init__.py
@@ -1,3 +1,3 @@
 from .dataset import Sequence
-from .upsampler import BatchUpsampler
+from .batch_upsampler import BatchUpsampler
 from .utils import get_sequence_or_none
